Add gline_lookup_badchan, replacing gline_find in badchan. This deals with the case where gline_find aborts on an inactive gline. gline_lookup for host glines deals with this case, gline_Find returns any gline.
In gline.c, always loop+call match over the channels. This deals with the case where an oper wishes to gline a channel containing a * or other similar character - as we only call find_channel, for a channel called #foo*, you would gline #foo\* which is not what find_channel expects. Albeit, this code will be more intensive then the previous version.